Fix distance calculation
The total distance was recalculated with the new DF on each revolution
instead of accumulating each delta distance per impulse. Furthermore
removed the rounding as it resulted in significant deviation.

+++ src/stroke.model.h
@@ -10,7 +10,7 @@ namespace StrokeModel
         unsigned int cleanDeltaTime;
         unsigned long rawRevTime;
         unsigned int driveDuration;
-        unsigned int distance;
+        double distance;
         double avgStrokePower;
         double dragCoefficient;
     };
